move helpers in test utils
